+++ pkg/innerring/config/config.go
@@ -40,8 +40,6 @@ type Config struct {
 
 	Workers Workers `mapstructure:"workers"`
 
-	Audit Audit `mapstructure:"audit"`
-
 	Indexer Indexer `mapstructure:"indexer"`
 
 	Contracts Contracts `mapstructure:"contracts"`
@@ -126,7 +124,6 @@ type Validator struct {
 // Applied only in debug mode.
 type Settlement struct {
 	BasicIncomeRate int64 `mapstructure:"basic_income_rate"`
-	AuditFee        int64 `mapstructure:"audit_fee"`
 }
 
 // Experimental configures experimental features.