+++ pkg/innerring/processors/settlement/events.go
@@ -4,27 +4,17 @@ import (
 	"github.com/nspcc-dev/neofs-node/pkg/morph/event"
 )
 
-// AuditEvent is an event of the start of
-// cash settlements for data audit.
-type AuditEvent struct {
+// BasicIncomeEvent is an event of the beginning of basic income cash settlements.
+type BasicIncomeEvent struct {
 	epoch uint64
 }
 
-type BasicIncomeEvent = AuditEvent
-
 // MorphEvent implements Neo: FS chain event.
-func (e AuditEvent) MorphEvent() {}
-
-// NewAuditEvent creates new AuditEvent for epoch.
-func NewAuditEvent(epoch uint64) event.Event {
-	return AuditEvent{
-		epoch: epoch,
-	}
-}
+func (e BasicIncomeEvent) MorphEvent() {}
 
 // Epoch returns the number of the epoch
 // in which the event was generated.
-func (e AuditEvent) Epoch() uint64 {
+func (e BasicIncomeEvent) Epoch() uint64 {
 	return e.epoch
 }